% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_phonology.R
\name{analyse_phonology}
\alias{analyse_phonology}
\title{Analyse data and create a report}
\usage{
analyse_phonology(correspondences, corpus, path)
}
\arguments{
\item{correspondences}{a checked dataframe with transcription-IPA correspondences}

\item{corpus}{variable containing corpus}

\item{path}{path, where function will create a report}
}
\description{
This function takes checked transcription-IPA correspondences and checked corpus or word list and generates a phonological report.
}
\examples{
# analyse_phonology(correspondences, corpus, path)
}
\author{
George Moroz <agricolamz@gmail.com>
}
