% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_converter.R
\name{apply_converter}
\alias{apply_converter}
\title{Apply converter to vector of strings}
\usage{
apply_converter(x, converter)
}
\arguments{
\item{correspondence}{a character vector, containing symbol correspondences to IPA symbols in orthographical system of user's files.}

\item{ipa}{a character vector, containing IPA symbols.}
}
\description{
This function returns a voctro with transcription and IPA symbols, simultaneously checking this dataframe for mistakes.
}
\examples{
apply_converter(zilo_dict$zilo_cyrillic, zilo_correspondences)
}
\author{
George Moroz <agricolamz@gmail.com>
}
