% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_converter.R
\name{check_converter}
\alias{check_converter}
\title{Check IPA-transcription correspondence}
\usage{
check_converter(correspondence, ipa)
}
\arguments{
\item{correspondence}{a character vector, containing symbol correspondences to IPA symbols in orthographical system of user's files.}

\item{ipa}{a character vector, containing IPA symbols.}
}
\description{
This function returns a dataframe with transcription and IPA symbols, simultaneously checking this dataframe for mistakes.
}
\examples{
check_converter(c("лъ", "и", "пI"), c("ɬ", "i", "pʼ"))
}
\author{
George Moroz <agricolamz@gmail.com>
}
