% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_sounds.R
\name{count_sounds}
\alias{count_sounds}
\title{Count sounds in a word list using converter}
\usage{
count_sounds(x, converter, context = 0)
}
\arguments{
\item{x}{a character vector with list of words for analysis.}

\item{converter}{a converter with transcription-IPA correspondences.}

\item{context}{integer value that define the context size.}
}
\description{
This function returns a dataframe with counts for all unique converted IPA combinations with context (if context argument greater then 0).
}
\examples{
count_sounds(zilo_dict$zilo_cyrillic, zilo_converter)
}
\author{
George Moroz <agricolamz@gmail.com>
}
