% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_correspondence.R
\name{create_correspondence}
\alias{create_correspondence}
\title{Create IPA-transcription correspondence}
\usage{
create_correspondence(correspondence = "")
}
\arguments{
\item{correspondence}{a character vector, containing symbol correspondences to IPA symbols in orthographical system of user's files}
}
\description{
This function returns a dataframe with IPA id, IPA symbol and transcription. If user doesn't provide any data, the function returns a template that is used in lingphonology, so users can save it to the disc and open it with appropriate software.
}
\examples{
create_correspondence()
}
\author{
George Moroz <agricolamz@gmail.com>
}
