% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_url.R
\name{create_url}
\alias{create_url}
\title{Create a link to PHOIBLE database}
\usage{
create_url(ipa)
}
\arguments{
\item{ipa}{a character vector, containing IPA symbols}
}
\description{
This function creates a to PHOIBLE database using a vector of IPA symbols.
}
\examples{
create_url(c("ɬ", "i", "pʼ"))
}
\author{
George Moroz <agricolamz@gmail.com>
}
