% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_conversion.R
\name{make_conversion}
\alias{make_conversion}
\title{Split and make conversion}
\usage{
make_conversion(x, converter)
}
\arguments{
\item{x}{character vector of words that should be converted.}

\item{converter}{correspondenses dataframe with columns "transcription" and "ipa"}
}
\description{
This is an auxiliary function, that takes a vector of strings for conversion. Returns a dataframe with notconverted vector, converted vector and convertions separated by space.
}
\author{
George Moroz <agricolamz@gmail.com>
}
