% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phonological_report.R
\docType{data}
\name{phonological_report}
\alias{phonological_report}
\title{Phonological report template}
\format{A chr vector with 29 rows.}
\usage{
phonological_report
}
\description{
Phonological report template
}
\keyword{datasets}
