% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zilo_converter.R
\docType{data}
\name{zilo_converter}
\alias{zilo_converter}
\title{Zilo transcription-IPA correspondences}
\format{A data frame with 138 rows and 2 variables:
\describe{
  \item{zilo_cyrillic}{Zilo words written in cyrillic script}
  \item{eng_translation}{English translations of the words}
  \item{pos}{Part of speach of the words}
}}
\usage{
zilo_converter
}
\description{
Zilo transcription-IPA correspondences based on \href{https://github.com/agricolamz/andi_cyrillic_to_IPA}{TECkit map} made by George Moroz and Neige Rochant.
}
\keyword{datasets}
