% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zilo_dict.R
\docType{data}
\name{zilo_dict}
\alias{zilo_dict}
\title{Zilo language dictionary}
\format{A data frame with 99 rows and 3 variables:
\describe{
  \item{zilo_cyrillic}{Zilo words written in cyrillic script}
  \item{eng_translation}{English translations of the words}
  \item{pos}{Part of speach of the words}
}}
\usage{
zilo_dict
}
\description{
Zilo language dictionary
}
\keyword{datasets}
